Move buffer functions out of server and into library
