+++ library/twopence.h
@@ -22,6 +22,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #define TWOPENCE_H
 
 #include <stdbool.h>
+#include "buffer.h"
 
 struct pollfd;
