+++ server/server.h
@@ -25,7 +25,6 @@
 #include "twopence.h"
 
 typedef struct socket socket_t;
-typedef struct twopence_buf twopence_buf_t;
 typedef struct packet packet_t;
 typedef struct queue queue_t;
 
@@ -72,37 +71,6 @@ struct header {
 #define PROTO_HDR_TYPE_MINOR	'm'
 #define PROTO_HDR_TYPE_TIMEOUT	'T'
 
-
-struct twopence_buf {
-	char *		base;
-	unsigned int	head;
-	unsigned int	tail;
-	unsigned int	size;
-};
-
-extern void		twopence_buf_init(twopence_buf_t *bp);
-extern void		twopence_buf_init_static(twopence_buf_t *bp, void *data, size_t len);
-extern void		twopence_buf_destroy(twopence_buf_t *bp);
-extern twopence_buf_t *	twopence_buf_new(size_t max_size);
-extern twopence_buf_t *	twopence_buf_clone(twopence_buf_t *bp);
-extern void		twopence_buf_free(twopence_buf_t *bp);
-extern const void *	twopence_buf_head(const twopence_buf_t *bp);
-extern void *		twopence_buf_tail(const twopence_buf_t *bp);
-extern unsigned int	twopence_buf_tailroom(const twopence_buf_t *bp);
-extern unsigned int	twopence_buf_tailroom_max(const twopence_buf_t *bp);
-extern unsigned int	twopence_buf_count(const twopence_buf_t *bp);
-extern void *		twopence_buf_pull(twopence_buf_t *bp, unsigned int len);
-extern bool		twopence_buf_push(twopence_buf_t *bp, void *data, unsigned int len);
-extern bool		twopence_buf_resize(twopence_buf_t *bp, unsigned int want_size);
-extern void		twopence_buf_reserve_head(twopence_buf_t *bp, unsigned int len);
-extern void *		twopence_buf_reserve_tail(twopence_buf_t *bp, unsigned int len);
-extern void		twopence_buf_advance_tail(twopence_buf_t *bp, unsigned int len);
-extern void		twopence_buf_advance_head(twopence_buf_t *bp, unsigned int len);
-extern void		twopence_buf_truncate(twopence_buf_t *bp, unsigned int len);
-extern bool		twopence_buf_append(twopence_buf_t *bp, const void *data, unsigned int len);
-extern bool		twopence_buf_puts(twopence_buf_t *bp, const char *s);
-extern void		twopence_buf_reset(twopence_buf_t *bp);
-extern void		twopence_buf_compact(twopence_buf_t *bp);
 
 extern packet_t *	packet_new(twopence_buf_t *bp);
 extern void		packet_free(packet_t *pkt);
